improving install script

+++ configure.sh
@@ -1,5 +1,48 @@
 #!/usr/bin/env bash
 
+install_prefix=""
+n_jobs=1
+
+while [[ $# -gt 0 ]]
+do
+    case $i in
+    --install-prefix)
+        install_prefix=$2
+        shift;shift
+        ;;
+    -n|--jobs)
+        n_jobs=$2
+        shift;shift
+        ;;
+    *)
+        echo
+        echo Usage:
+        echo '--install-prefix <path>'
+        echo '-n|--jobs <int>'
+        echo
+        ;;
+    esac
+done
+
+[ -z $install_prefix ] && {
+    read -p 'No install prefix set. Set prefix? [yn] ' y
+    if [ "$y" == y ]
+    then
+        read -e -p "Enter prefix: " tmp
+        install_prefix=${tmp/\~/$HOME}
+    else
+        install_prefix="$HOME/.local"
+    fi
+}
+
+
+realpath $install_prefix || {
+    echo "Instal prefix $install_prefix not found."
+    exit 1
+}
+install_prefix=$(realpath $install_prefix)
+echo "Using install prefix $install_prefix"
+
 cp vimrc_base ~/.vimrc
 [ -d ~/.vim ] || mkdir ~/.vim
 
@@ -22,10 +65,65 @@ do
     cp $i ~/.vim/$i
 done
 
+read -p 'Install Dash? [yn] ' inst
+if [ "$inst" == "y" ]
+then
+    echo
+    echo 'Installing dash from source...'
+    echo
+    wget http://gondor.apana.org.au/~herbert/dash/files/dash-0.5.10.2.tar.gz \
+        -O external/dash.tar.gz
+    pushd external
+    tar -gvxf dash.tar.gz
+    pushd $(ls | grep dash | grep -v tar)
+
+    read -p "Install dash to $install_prefix? [yn] " y
+    if [ "$y" == "y" ]
+    then
+        ./configure --bindir=$install_prefix/bin --mandir=$install_prefix/share/man || \
+        {
+            echo Error configuring dash installation.
+            exit 1
+        }
+        make -j $n_jobs || \
+        {
+            echo Error making dash...
+            exit 1
+        }
+        make install || \
+        {
+            echo Error installing dash...
+            exit 1
+        }
+        echo
+        echo 'Dash installed'
+        echo
+    else
+        ./configure || \
+        {
+            echo Error configuring dash installation.
+            exit 1
+        }
+        make -j $n_jobs || \
+        {
+            echo Error making dash...
+            exit 1
+        }
+        echo
+        echo 'Dash built'
+        echo
+    fi
+    popd; popd
+else
+    echo
+    echo 'Skipping dash installation'
+    echo
+fi
+
+
 type tmux
 if [ $? -ne 0 ]; then
-    echo 'Install tmux? [yn]'
-    read inst
+    read -p 'Install tmux? [yn] ' inst
     if [ "$inst" == "y" ]; then
         echo
         echo Installing tmux from tarball...
@@ -36,10 +134,10 @@ if [ $? -ne 0 ]; then
         tar -xvgf ./external/tmux-3.0.tar.gz
         pushd tmux-3.0
         ./configure \
-            --prefix=$HOME/.local \
-            CPPFLAGS="-I$HOME/.local/include" \
-            LDFLAGS="-L$HOME/.local/lib"
-        make -j 8
+            --prefix=$install_prefix \
+            CPPFLAGS="-I$install_prefix/include" \
+            LDFLAGS="-L$install_prefix/lib"
+        make -j $n_jobs
         make install
         popd
         popd
@@ -53,21 +151,24 @@ cp tmux.conf $HOME/.tmux.conf
 
 which zsh
 if [ $? -ne 0 ]; then
-    echo
-    echo Installing Zsh from source...
-    echo
-    pushd external
-    wget -O zsh.tar.xz https://sourceforge.net/projects/zsh/files/latest/download
-    [ -d zsh ] || mkdir zsh
-    tar xf zsh.tar.xz -C zsh --strip-components 1
-    pushd zsh
-    ./configure \
-        --prefix=$HOME/.local \
-        CPPFLAGS="-I$HOME/.local/include" \
-        LDFLAGS="-L$HOME/.local/lib"
-    make -j 8
-    make install
-    popd; popd
+    read -p 'Install zsh?' inst
+    if [ "$inst" == "y" ]; then
+        echo
+        echo Installing Zsh from source...
+        echo
+        pushd external
+        wget -O zsh.tar.xz https://sourceforge.net/projects/zsh/files/latest/download
+        [ -d zsh ] || mkdir zsh
+        tar xf zsh.tar.xz -C zsh --strip-components 1
+        pushd zsh
+        ./configure \
+            --prefix=$install_prefix \
+            CPPFLAGS="-I$install_prefix/include" \
+            LDFLAGS="-L$install_prefix/lib"
+        make -j $n_jobs
+        make install
+        popd; popd
+    fi
 fi
 
 if [ ! -d $HOME/.oh-my-zsh/plugins/zsh-syntax-highlighting ]
@@ -83,8 +184,7 @@ if [ $? -ne 0 ]; then
     echo Ctags not found...
     echo
 else
-    echo 'Generate new ctags? [yn]'
-    read inst
+    read -p 'Generate new ctags? [yn] ' inst
     if [ "$inst" == "y" ]; then
         [ -f "./tags/$(uname -n)" ] || {
             echo
@@ -95,12 +195,12 @@ else
         echo
         echo Generating Ctags...
         echo
-        [ -f ./tags ] && rm ./tags
-        touch ./tags
+        [ -f ./tags-file ] && rm ./tags-file
+        touch ./tags-file
         while read pth
         do
             find $pth -name '*' | grep -E '\.(h|hpp)$' | \
-                 ctags -f ./tags \
+                 ctags -f ./tags-file \
                  --verbose \
                  --append \
                  --c++-kinds=+p \
@@ -110,7 +210,7 @@ else
                  -I _GLIBCXX_NOEXCEPT \
                  -L -
         done < "./tags/$(uname -n)"
-        mv ./tags ~/.vim/tags
+        mv ./tags-file ~/.vim/tags
         echo "set tags=$(realpath $HOME/.vim/tags),tags;" >> ~/.vimrc
     fi
 fi
